feat(scripts): use snapshots for droplet
This will cost more, but will make devserver more useful.

When destroying the droplet, the client will take a snapshot of the
droplet, delete any previous snapshots, and then destroy the droplet.

When creating the droplet, the client will check if a snapshot exists.
If a snapshot exists, a new droplet will be created from it. Otherwise,
a new droplet will be created from the docker image.